+++ mordant/src/appleMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.apple.kt
@@ -1,3 +1,5 @@
+@file:OptIn(ExperimentalForeignApi::class)
+
 package com.github.ajalt.mordant.internal
 
 import kotlinx.cinterop.ExperimentalForeignApi
@@ -8,7 +10,6 @@ import platform.posix.TIOCGWINSZ
 import platform.posix.ioctl
 import platform.posix.winsize
 
-@OptIn(ExperimentalForeignApi::class)
 internal actual fun getTerminalSize(): Size? {
     return memScoped {
         val size = alloc<winsize>()