+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.kt
@@ -56,3 +56,9 @@ internal expect fun sendInterceptedPrintRequest(
 internal expect val FAST_ISATTY: Boolean
 
 internal expect val CR_IMPLIES_LF: Boolean
+
+internal expect fun exitProcessMpp(status: Int)
+
+internal expect fun readFileIfExists(filename: String): String?
+
+internal expect fun hasFileSystem(): Boolean