+++ mordant/src/jsMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.js.kt
@@ -19,6 +19,12 @@ private class NodeMppImpls(private val fs: dynamic) : BaseNodeMppImpls<dynamic>(
     override fun stdoutInteractive(): Boolean = js("Boolean(process.stdout.isTTY)") as Boolean
     override fun stdinInteractive(): Boolean = js("Boolean(process.stdin.isTTY)") as Boolean
     override fun stderrInteractive(): Boolean = js("Boolean(process.stderr.isTTY)") as Boolean
+    override fun exitProcess(status: Int) {
+        process.exit(status)
+    }
+    override fun cwd(): String {
+        return process.cwd() as String
+    }
     override fun getTerminalSize(): Size? {
         // For some undocumented reason, getWindowSize is undefined sometimes, presumably when isTTY
         // is false
@@ -41,6 +47,10 @@ private class NodeMppImpls(private val fs: dynamic) : BaseNodeMppImpls<dynamic>(
 
     override fun makeTerminalCursor(terminal: Terminal): TerminalCursor {
         return NodeTerminalCursor(terminal)
+    }
+
+    override fun readFileIfExists(filename: String): String? {
+        return fs.readFileSync(filename, "utf-8") as? String
     }
 }