+++ mordant/src/macosMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.macos.kt
@@ -1,0 +1,3 @@
+package com.github.ajalt.mordant.internal
+
+internal actual fun hasFileSystem(): Boolean = true