+++ mordant/src/mingwMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.mingw.kt
@@ -39,3 +39,5 @@ internal actual fun ttySetEcho(echo: Boolean) = memScoped {
     }
     SetConsoleMode(stdinHandle, newMode)
 }
+
+internal actual fun hasFileSystem(): Boolean = true